pyproject.toml
[tool.poetry]
name = "kunitori"
version = "0.1.0"
description = ""
authors = ["Yuuki Takahashi <20282867+yktakaha4@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.24"
Jinja2 = "^3.0.3"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kunitori = "kunitori.cli:main"
